requirements.txt
Django==5.2.8
python-decouple==3.8
requests==2.32.3
redis==5.0.8
django-redis==5.4.0
Pillow==10.4.0
